Fix improper use of result
The fetchOne method used in the model.zep in the model's exists method
returns an object if the pdo connection has a default return type set
to object.

The model exists method expects the return to be an associative array.
If the model get an object in the altered method it will throw a notice
and assume the row does not yet exit.  It will later on atempt an
instert when a update is actually needed.

This fix insures that the model always gets an associative array in the
exists method.